# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.LoginRequestDto:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.LoginResponseDto:
    properties:
      accessToken:
        type: string
      userId:
        type: string
    type: object
  dto.RekeningDtoSwag:
    properties:
      rekening:
        type: string
    type: object
...
  dto.UpdatePinRequestSwag:
    properties:
      new_pin:
        type: string
      old_pin:
        type: string
    required:
    - new_pin
    - old_pin
    type: object
  dto.UpdatePinResponse:
    properties:
      pin:
        type: string
      user_id:
        type: string
    type: object
  dto.UserRequestDto:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  dto.UserRequestEditDto:
    properties:
      email:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
  dto.VerifyUser:
    properties:
      jenis_kelamin:
        type: string
      nik:
        type: string
      photo:
        type: string
      pin:
        type: string
      tanggal_lahir:
        type: string
      umur:
        type: integer
      user_id:
        type: string
    type: object
...
  model.Rekening:
    properties:
      created_at:
        type: string
      id:
        type: string
      rekening:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      role:
        type: string
      saldo:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  model.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      name:
        type: string
      phone_number:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
  model.UserSaldo:
    properties:
      pin:
        type: string
      saldo:
        type: integer
      user:
        $ref: '#/definitions/model.UserResponse'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: A tag service API in Go using Gin framework
  title: Tag Service API
  version: "1.0"
paths:
...
  /users:
    post:
      description: Return the info user
      parameters:
      - description: Create user
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/dto.UserRequestDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.User'
      summary: User register account.
      tags:
      - User
    put:
      description: Return new the info user
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Edit form
        in: body
        name: Edit
        required: true
        schema:
          $ref: '#/definitions/dto.UserRequestEditDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
      security:
      - ApiKeyAuth: []
      summary: User edit account
      tags:
      - User
  /users/{id}:
    get:
      description: Return the balance user with jwt id
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserSaldo'
      security:
      - ApiKeyAuth: []
      summary: admin Get balance user with id param.
      tags:
      - Admin
  /users/login:
    post:
      description: Return the access token and id user
      parameters:
      - description: Login form
        in: body
        name: Login
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponseDto'
      summary: User login to get jwtAuth
      tags:
      - User
  /users/pin:
    put:
      consumes:
      - application/json
      description: Update user PIN with JSON payload
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Update PIN Request Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePinRequestSwag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdatePinResponse'
      security:
      - ApiKeyAuth: []
      summary: Update user PIN
      tags:
      - User
  /users/rekening:
    get:
      consumes:
      - application/json
      description: Get Rekening user
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Rekening'
      security:
      - ApiKeyAuth: []
      summary: Get Rekening User
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create Rekening user to withdraw
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Rekening data
        in: body
        name: rekening
        required: true
        schema:
          $ref: '#/definitions/dto.RekeningDtoSwag'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Rekening'
      security:
      - ApiKeyAuth: []
      summary: Create Rekening User
      tags:
      - User
  /users/saldo:
    get:
      description: Return the info user and balance
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserSaldo'
      security:
      - ApiKeyAuth: []
      summary: User Get their info balance
      tags:
      - User
  /users/verify:
    post:
      consumes:
      - multipart/form-data
      description: Return info user data
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: JSON data of user
        in: formData
        name: user
        required: true
        type: string
      - description: Photo
        in: formData
        name: photo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.VerifyUser'
      security:
      - ApiKeyAuth: []
      summary: User verify account
      tags:
      - User
swagger: "2.0"